+++ src/xsdba/adjustment.py
@@ -534,7 +534,7 @@ class DetrendedQuantileMapping(TrainAdjust):
     adapt_freq_thresh : str, optional
         Threshold for frequency adaptation. See :py:class:`xsdba.processing.adapt_freq` for details.
         Default is None, meaning that frequency adaptation is not performed.
-    n_last_quantile_filter: int, optional
+    n_last_quantile_filter: float, optional
         If not None, values to adjust (after preprossing steps) that are above n_last_quantile_filter * the value
         of the last quantile of hist (before the preprocessing steps, stored in hist_q_raw) are not adjusted.
         We keep the input simulation with only the preprocessing steps instead.
@@ -590,7 +590,7 @@ class DetrendedQuantileMapping(TrainAdjust):
         jitter_under_thresh_value: str | None = None,
         jitter_over_thresh_value: str | None = None,
         jitter_over_thresh_upper_bnd: str | None = None,
-        n_last_quantile_filter: int | None = None,
+        n_last_quantile_filter: float | None = None,
     ):
         if group.prop not in ["group", "dayofyear"]:
             warn(f"Using DQM with a grouping other than 'dayofyear' is not recommended (received {group.name}).", stacklevel=2)
